pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "priorcons"
version = "0.2.1"
description = "Tool for the integration of viral consensus sequences obtained by de novo and mapping strategies, supported by prior information."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Germán Vallejo Palma", email = "german.vallejo@isciii.es" }
]
requires-python = ">=3.8"
dependencies = [
  "biopython>=1.79",
  "numpy>=1.21",
  "pandas>=1.3",
  "pyarrow>=12.0"  
]

[project.scripts]
priorcons = "priorcons.cli:main"

[project.urls]
Homepage = "https://github.com/GERMAN00VP/priorCons/"
